changed index from int to size_t

+++ ft_itoa.c
@@ -37,7 +37,7 @@ static int	ft_getdigit(unsigned int n)
 	return (count);
 }
 
-static char	*ft_convertint(char *str, int index, unsigned int n, int digit)
+static char	*ft_convertint(char *str, size_t index, unsigned int n, int digit)
 {
 	while (1 < digit)
 	{
@@ -54,7 +54,7 @@ char	*ft_itoa(int n)
 {
 	int				digit;
 	char			*str;
-	int				index;
+	size_t			index;
 	unsigned int	num;
 
 	index = 0;
